Cleaning Up The Lexor Fixes #73
